split up feed event comment
