+++ apps/mobile/src/components/Feed/Socialize/CommentBox.tsx
@@ -8,16 +8,11 @@ import Animated, {
   useSharedValue,
   withSpring,
 } from 'react-native-reanimated';
-import { ConnectionHandler, fetchQuery, graphql, useRelayEnvironment } from 'react-relay';
-import { SelectorStoreUpdater } from 'relay-runtime';
+import { useEventComment } from 'src/hooks/useEventComment';
 import { XMarkIcon } from 'src/icons/XMarkIcon';
 import useKeyboardStatus from 'src/utils/useKeyboardStatus';
 
 import { GalleryTouchableOpacity } from '~/components/GalleryTouchableOpacity';
-import { CommentBoxMutation } from '~/generated/CommentBoxMutation.graphql';
-import { CommentBoxQuery } from '~/generated/CommentBoxQuery.graphql';
-import { useReportError } from '~/shared/contexts/ErrorReportingContext';
-import { usePromisifiedMutation } from '~/shared/relay/usePromisifiedMutation';
 import colors from '~/shared/theme/colors';
 
 import { SendIcon } from './SendIcon';
@@ -33,11 +28,12 @@ type Props = {
 };
 
 export function CommentBox({ autoFocus, onClose, isNotesModal = false, feedEventId }: Props) {
-  const reportError = useReportError();
   const { colorScheme } = useColorScheme();
   const [value, setValue] = useState('');
 
   const characterCount = useMemo(() => 100 - value.length, [value]);
+
+  const { submitComment, isSubmittingComment } = useEventComment();
 
   const isKeyboardActive = useKeyboardStatus();
 
@@ -57,120 +53,19 @@ export function CommentBox({ autoFocus, onClose, isNotesModal = false, feedEvent
     return isKeyboardActive;
   }, [isKeyboardActive, isNotesModal]);
 
-  const [submitComment, isSubmittingComment] = usePromisifiedMutation<CommentBoxMutation>(graphql`
-    mutation CommentBoxMutation($eventId: DBID!, $comment: String!, $connections: [ID!]!)
-    @raw_response_type {
-      commentOnFeedEvent(comment: $comment, feedEventId: $eventId) {
-        ... on CommentOnFeedEventPayload {
-          __typename
-
-          comment @appendNode(connections: $connections, edgeTypeName: "FeedEventCommentEdge") {
-            dbid
-            __typename
-            comment
-            commenter {
-              dbid
-              id
-              username
-            }
-            creationTime
-          }
-        }
-      }
-    }
-  `);
-
-  const relayEnvironment = useRelayEnvironment();
-  const handleSubmit = useCallback(async () => {
+  const handleSubmit = useCallback(() => {
     if (value.length === 0) {
       return;
     }
 
-    try {
-      // We're using `fetchQuery` here instead of a `useFragment` in the render method
-      // since we don't need our data requirements blocking the initial render of the screen.
-      // We also don't have to prop drill a bunch of refs down to get what we need.
-      // If the data you need is only needed in a callback, it's fine to use `fetchQuery`.
-      const eventRelayId = `FeedEvent:${feedEventId}`;
-      const query = await fetchQuery<CommentBoxQuery>(
-        relayEnvironment,
-        graphql`
-          query CommentBoxQuery {
-            viewer {
-              ... on Viewer {
-                user {
-                  id
-                  dbid
-                  username
-                }
-              }
-            }
-          }
-        `,
-        {},
-        { fetchPolicy: 'store-or-network' }
-      ).toPromise();
-
-      const interactionsConnection = ConnectionHandler.getConnectionID(
-        eventRelayId,
-        'Interactions_comments'
-      );
-      const commentsBottomSheetConnection = ConnectionHandler.getConnectionID(
-        eventRelayId,
-        'CommentsBottomSheet_comments'
-      );
-
-      const updater: SelectorStoreUpdater<CommentBoxMutation['response']> = (store, response) => {
-        if (response.commentOnFeedEvent?.__typename === 'CommentOnFeedEventPayload') {
-          const pageInfo = store.get(interactionsConnection)?.getLinkedRecord('pageInfo');
-
-          pageInfo?.setValue(((pageInfo?.getValue('total') as number) ?? 0) + 1, 'total');
-        }
-      };
-
-      const optimisticId = Math.random().toString();
-      const response = await submitComment({
-        updater,
-        optimisticUpdater: updater,
-        optimisticResponse: {
-          commentOnFeedEvent: {
-            __typename: 'CommentOnFeedEventPayload',
-            comment: {
-              __typename: 'Comment',
-              comment: value,
-              commenter: {
-                dbid: query?.viewer?.user?.dbid ?? 'unknown',
-                id: query?.viewer?.user?.id ?? 'unknown',
-                username: query?.viewer?.user?.username ?? null,
-              },
-              creationTime: new Date().toISOString(),
-              dbid: optimisticId,
-              id: `Comment:${optimisticId}`,
-            },
-          },
-        },
-        variables: {
-          comment: value,
-          eventId: feedEventId,
-          connections: [interactionsConnection, commentsBottomSheetConnection],
-        },
-      });
-
-      if (response.commentOnFeedEvent?.__typename === 'CommentOnFeedEventPayload') {
+    submitComment({
+      value,
+      feedEventId,
+      onSuccess: () => {
         resetComment();
-      } else {
-        reportError(
-          `Error while commenting on feed event, typename was ${response.commentOnFeedEvent?.__typename}`
-        );
-      }
-    } catch (error) {
-      if (error instanceof Error) {
-        reportError(error);
-      } else {
-        reportError('An unexpected error occurred while posting a comment.');
-      }
-    }
-  }, [value, feedEventId, relayEnvironment, submitComment, resetComment, reportError]);
+      },
+    });
+  }, [feedEventId, resetComment, submitComment, value]);
 
   const disabledSendButton = useMemo(() => {
     return value.length === 0 || characterCount < 0 || isSubmittingComment;
